Creating bands objects (#83)
This PR improves how the code stores information associated with the orbitals/bands. It does so by introducing the `Band` and `Bands` classes.

This PR does *not* alter the behaviour of the code or introduce any new features. All it does is simplify & tidy the underlying code (e.g. see the diff of `kc_with_cp.py`). These changes should make it cleaner to implement orbital tracking.