Changing from https to ssh for submodules (#75)
The test suite previously required the submodules to be linked via `https` in order to function. This is no longer the case.

Now, the test suite uses the git user `koopmans-tester`. I set up github and gitlab accounts for `koopmans-tester`. I have given it read-only access to this repository, and its email and private key are stored as secrets. The introduction of this git user means that the test suite can now checkout the repository using `ssh`.

This change means that we are no longer tied down to using https for the submodules more generally. Now, code developers will not have to regularly enter passwords etc. in order to push to submodules linked via https -- instead, they can simply use `ssh`