Unifying workflow creation (#122)
Previously, when we read in a JSON file there was a lot of logic that helped auto-generate keywords. However, this same logic was not invoked when directly instantiating a workflow.

This PR fixes this issue. As much of the logic of `io._json.read_json()` has been moved into the `Workflow` class. This class now has a `.fromjson()` `@classmethod` which allows you to create a workflow from a json file (although `wf = io.read("filename.json"`) still works)