Fix write json and new keyword kpath_density (#125)
### Writing json files
The function for writing json input files was outdated and incapable of writing full output files.

This PR fixes that. Now, we should be able to write any workflow to a correct json file. (This resolves #115.)

This PR also sees the introduction of some new test scripts such as `koopmans/io/test_io.py`. This is the model we will eventually move to for the test suite (with much shorter, targeted tests) but currently these tests to not get run by the automated suite. You can run these tests by running `pytest koopmans/`

### kpath_density
In an unrelated change, this PR also introduces the new keyword `kpath_density` to the `k_points` block. This allows the user to specify the number of points they would like along their kpath (which up until now they were stuck with the ASE default). This resolves #95