Generalising enforce spin symmetry (#127)
In many situations, we would like to avoid spin contamination. The way we do this is first run a spin-1 calculation, duplicate the wave function to generate a spin-2, spin-symmetric wave function, and then restart from this point a spin 2 calculation.

However, the current code for performing this procedure was specific to `kcp.x` calculations. This is in spite of the fact that this procedure was implemented inside of `Workflow.run_calculator()`, which -- being a method of a `Workflow` object -- should be agnostic to the kind of calculator it receives.

This PR rearranges the code so it is a lot more logical. Now, calculators that can have spin symmetry enforced inherit from the abstract base class `CalculatorCanEnforceSpinSym`, and `Workflow.run_calculator()` only uses the methods defined in this ABC. Any calculator-specific logic is contained within that calculator as it defines the various functions set out by `CalculatorCanEnforceSpinSym`.

Currently, only `kcp.x` calculators have this functionality actually implemented. In the future we could consider making the `pw.x` calculator inherit from `CalculatorCanEnforceSpinSym` if we encounter issues with spin contamination from that code.

Closes #93