Added the ability to calculate projected DOS (#106)
pDOS calculation is now supported. The pDOS will be generated if...
- a wannierization or dft_bands task is being performed
- the pseudos being used have the required `PP_PSWFC` entries
- the degauss is suitably large compared to the density of the sampling of the k-path (in the case of UI bandstructures)

We have also unified the plotting of bandstructures and DOSs. This is now performed with the function `workflow.plot_bandstructure(bs, dos, ...)` where if `dos` is not None it plots additionally the DOS alongside. Customisation of the plot can be performed via various keyword arguments. See `Workflow.plot_bandstructure(...)` for details.

Other minor changes:
- fixed error in the QE version tag `koopmans` was printing out
- fixed the printing of the "koopmans" header so it always appears before warnings, and appears when workflows are run from within python (which it previously did not)