Temporary directory autoremoval (#148)
This PR adds the option `keep_tmpdirs` to the `workflow` block. If set to `False`, then the workflow, once finished, will delete all the `outdirs` from the calculations it performed. `keep_tmpdirs` is `True` by default.

This PR also changes the default value of `from_scratch` from `False` to `True`, because I think this is better choice for the default.

Unrelatedly, this PR also moves some old test routines from `koopmans/` into `tests/`, which means they are now run as part of the test suite.