Converting to US spelling (#157)
Ultimately opted to make all of the code base use US spelling. While I prefer UK spelling, insisting on using UK spelling leads to strange code when interacting with python libraries which almost exclusively use US. Note that the one place that we need to be careful is for `wannier90`, which uses UK spelling. This closes #92 

Other changes:
- added a check that the docs successfully build, which caught a few minor errors in the docs that were then fixed
- added a pre-commit configuration so that we all use the same pre-commit