Fixing ASE autodoc (#166)
Minor change to the docs `conf.py` file and the readthedocs config to make ``autodoc`` work.

Specifically, ``autodoc`` was not working because it did not know about ``ase``. It has been added as a ``autodoc_mock_import``. Meanwhile, the requirements for ``koopmans`` have also been added as requirements for the docs.