Default nrb values (#151)
The dimension of  the small box `nr1b`, `nr2b` and `nr3b` now have values that are automatically generated by the code, instead of relying on the user to specify them.

Ideally this would be done using the cutoff radius (`rc`) of the pseudopotentials. Unfortunately the pseudopotentials do not store this information in a standardized way.

The solution for now is to set it to a safe value rc=3.0 a.u. The largest rc in the DOJO pseudos is 2.6.

Closes #71