Performing DFPT on a coarser grid (#194)
This PR implements the ability to calculate the screening parameters via DFPT on a coarser grid than that which is used to construct the Hamiltonian. This is controlled by the keyword `dfpt_coarse_grid` in the `workflow` block, which is a direct analogue of the `grid` keyword in the `kpoints` block.

It also introduces the ability to group variational orbitals based on their spread (provided that they are Wannier functions). This is required in order to perform orbital grouping for DFPT workflows.

Finally, this PR makes `koopmans` suppress traceback by default. This can be re-enabled by running `koopmans` with the flag `--traceback` or `-t`.