Updated license to GPLv2 (#208)
N.B. any program containing code that is GPLv2 must itself be GPL v2. QE is GPL v2 (although it is distributed as a submodule so it is not clear if we are really "containing" the QE code or simply redistributing it separately under its own license. Nevertheless, better safe than sorry!