Adding version information to kwf files (#212)
Currently `.kwf` files can be quite fickle because the format relies on the structure of the `Workflow` object itself, so whenever we change the `Workflow` class we lose the ability to read old `kwf` files.

Now, we have added explicitly the version number to the `Workflow` class. This will warn users if they are loading a `.kwf` file that was generated with a different version of `koopmans` than that which generated it, and, if the reading fails, they can always revert to the version of `koopmans` that generated the file in the first place.