Providing default metadata for w90
the custom metadata can be provided in input json, as in examples/silicon_dfpt/si.json

trying also to delete all the AiiDA class nodes and information from the DFPTWorkflow class,
because it gives problems when in the end it tries to json serialize all the attributes.
The issue shows up only if we run from cmd line. Indeed the serialization is called only in that case.