Refactor to run all calculations in separate directories (#220)
Summary
-------
This refactor of the code contains several changes designed to make `koopmans` more robust and amenable to integration with `AiiDA`. Namely...
- each calculation is run in a separate directory following a regular pattern
- there are no shared `tmp` directories
- file reading/writing/moving is done in an abstract way (see the new `FilePointer` class) so that it will be possible to "move" files on a remote server
- the introduction of a `Process` class to make python operations on files etc. more standardized (and able to be executed remotely)

These changes to treat calculations more like pure functions, and the introduction of a `Process` class also starts us in the direction of following [CWL](https://www.commonwl.org/)'s design pattern more closely. In the long term, we would like to be as close as possible to CWL (perhaps even with composite workflows being able to be written in CWL).

Other changes
---------------------
- the machine-learning workflows have been simplified
- the output of `koopmans` markdown-compliant, making the output files easier to read for humans
- `.kwf` files have been replaced by `.pkl` files (removing the responsibility of writing `Workflows` to file from custom code and using the widely-used `dill` package instead)