Added support for different Engines (#235)
As part of interfacing `koopmans` with `AiiDA`, this PR shifts the responsibility of running calculations, file manipulation, etc to the `Engine`, a new abstract base class. `koopmans` currently ships with a `LocalhostEngine`, which runs calculations locally (i.e. the original behaviour of `koopmans`).

In order to achieve this, workflow.run() contains a while loop within which the workflow is repetitively executed. Whenever workflows encounter a calculation to perform, they submit it and then if the calculation status is not `COMPLETED`, then the workflow returns to the while loop.
- A `LocalhostEngine` runs the calculation locally upon submission, so the status should always be `COMPLETED` and the workflow will never traverse the while loop multiple times
- Other implementations (e.g. the `AiiDAEngine`) will enter the while loop while waiting for calculations to complete